Create File Search from one folder to another.
This code pick file from one folder and search into another folder. if file is not present in the folder the it write the name and id of the file in the sheet. this code search only docs type file as mime type is given u can change mime type and search different types of file in folder.